wheels_/backend/requirements.txt
# WHEELS Matchmaking API Dependencies
flask==2.3.3
flask-cors==4.0.0
supabase==2.18.1
pandas==2.0.3
geopy==2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
